feat: New Field mixins
