+++ src/Form/FieldClass.php
@@ -4,7 +4,6 @@ namespace Kirby\Form;
 
 use Kirby\Cms\HasSiblings;
 use Kirby\Toolkit\I18n;
-use Kirby\Toolkit\Str;
 
 /**
  * Abstract field class to be used instead
@@ -22,24 +21,24 @@ use Kirby\Toolkit\Str;
 abstract class FieldClass
 {
 	use HasSiblings;
+	use Mixin\After;
 	use Mixin\Api;
+	use Mixin\Autofocus;
+	use Mixin\Before;
+	use Mixin\Help;
+	use Mixin\Icon;
+	use Mixin\Label;
 	use Mixin\Model;
+	use Mixin\Placeholder;
 	use Mixin\Translatable;
 	use Mixin\Validation;
 	use Mixin\Value;
 	use Mixin\When;
+	use Mixin\Width;
 
-	protected string|null $after;
-	protected bool $autofocus;
-	protected string|null $before;
 	protected bool $disabled;
-	protected string|null $help;
-	protected string|null $icon;
-	protected string|null $label;
 	protected string|null $name;
-	protected string|null $placeholder;
 	protected Fields $siblings;
-	protected string|null $width;
 
 	public function __construct(
 		protected array $params = []
@@ -75,21 +74,6 @@ abstract class FieldClass
 		return $this->params[$param] ?? null;
 	}
 
-	public function after(): string|null
-	{
-		return $this->stringTemplate($this->after);
-	}
-
-	public function autofocus(): bool
-	{
-		return $this->autofocus;
-	}
-
-	public function before(): string|null
-	{
-		return $this->stringTemplate($this->before);
-	}
-
 	/**
 	 * Returns optional dialog routes for the field
 	 */
@@ -114,31 +98,9 @@ abstract class FieldClass
 		return [];
 	}
 
-	/**
-	 * Optional help text below the field
-	 */
-	public function help(): string|null
-	{
-		if (empty($this->help) === false) {
-			$help = $this->stringTemplate($this->help);
-			$help = $this->kirby()->kirbytext($help);
-			return $help;
-		}
-
-		return null;
-	}
-
 	protected function i18n(string|array|null $param = null): string|null
 	{
 		return empty($param) === false ? I18n::translate($param, $param) : null;
-	}
-
-	/**
-	 * Optional icon that will be shown at the end of the field
-	 */
-	public function icon(): string|null
-	{
-		return $this->icon;
 	}
 
 	public function id(): string
@@ -157,16 +119,6 @@ abstract class FieldClass
 	}
 
 	/**
-	 * The field label can be set as string or associative array with translations
-	 */
-	public function label(): string
-	{
-		return $this->stringTemplate(
-			$this->label ?? Str::ucfirst($this->name())
-		);
-	}
-
-	/**
 	 * Returns the field name
 	 */
 	public function name(): string
@@ -180,14 +132,6 @@ abstract class FieldClass
 	public function params(): array
 	{
 		return $this->params;
-	}
-
-	/**
-	 * Optional placeholder value that will be shown when the field is empty
-	 */
-	public function placeholder(): string|null
-	{
-		return $this->stringTemplate($this->placeholder);
 	}
 
 	/**
@@ -217,39 +161,9 @@ abstract class FieldClass
 		];
 	}
 
-	protected function setAfter(array|string|null $after = null): void
-	{
-		$this->after = $this->i18n($after);
-	}
-
-	protected function setAutofocus(bool $autofocus = false): void
-	{
-		$this->autofocus = $autofocus;
-	}
-
-	protected function setBefore(array|string|null $before = null): void
-	{
-		$this->before = $this->i18n($before);
-	}
-
 	protected function setDisabled(bool $disabled = false): void
 	{
 		$this->disabled = $disabled;
-	}
-
-	protected function setHelp(array|string|null $help = null): void
-	{
-		$this->help = $this->i18n($help);
-	}
-
-	protected function setIcon(string|null $icon = null): void
-	{
-		$this->icon = $icon;
-	}
-
-	protected function setLabel(array|string|null $label = null): void
-	{
-		$this->label = $this->i18n($label);
 	}
 
 	protected function setName(string|null $name = null): void
@@ -257,27 +171,11 @@ abstract class FieldClass
 		$this->name = strtolower($name ?? $this->type());
 	}
 
-	protected function setPlaceholder(array|string|null $placeholder = null): void
-	{
-		$this->placeholder = $this->i18n($placeholder);
-	}
-
 	protected function setSiblings(Fields|null $siblings = null): void
 	{
 		$this->siblings = $siblings ?? new Fields([$this]);
 	}
 
-	/**
-	 * Setter for the field width
-	 */
-	protected function setWidth(string|null $width = null): void
-	{
-		$this->width = $width;
-	}
-
-	/**
-	 * Returns all sibling fields for the HasSiblings trait
-	 */
 	protected function siblingsCollection(): Fields
 	{
 		return $this->siblings;
@@ -313,14 +211,5 @@ abstract class FieldClass
 	public function type(): string
 	{
 		return lcfirst(basename(str_replace(['\\', 'Field'], ['/', ''], static::class)));
-	}
-
-	/**
-	 * Returns the width of the field in
-	 * the Panel grid
-	 */
-	public function width(): string
-	{
-		return $this->width ?? '1/1';
 	}
 }
